[#50] 브랜치 변경, controller분할, 이름변경
