--- src/main/java/com/zerobase/munbanggu/study/model/entity/Study.java
@@ -1,78 +1,0 @@
-package com.zerobase.munbanggu.study.model.entity;
-
-
-import com.zerobase.munbanggu.user.type.EnrollmentStatus;
-import com.zerobase.munbanggu.user.type.RefundCycle;
-import java.time.LocalDateTime;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.jetbrains.annotations.Nullable;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Study {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String user_id;
-
-    private String title;
-
-    private String content;
-
-    private long min_user;
-
-    private long max_user;
-
-    private boolean public_or_not;
-
-    @Nullable
-    private String password;
-
-    private LocalDateTime start_date;
-
-    private LocalDateTime end_date;
-
-    @Nullable
-    private boolean start_rule;
-
-    @Nullable
-    private boolean start_attend_or_not;
-
-    private long checklist_cycle;
-    private long fee;
-
-    @Enumerated(EnumType.STRING)
-    @Nullable
-    private RefundCycle refundCycle;
-
-    @Enumerated(EnumType.STRING)
-    private EnrollmentStatus status;
-
-    @CreatedDate
-    private LocalDateTime create_date;
-
-    @Nullable
-    private LocalDateTime delete_date;
-
-
-
-}
