+++ src/main/java/com/zerobase/munbanggu/user/service/UserService.java
@@ -4,7 +4,7 @@ package com.zerobase.munbanggu.user.service;
 import static com.zerobase.munbanggu.type.ErrorCode.INVALID_CODE;
 import static com.zerobase.munbanggu.type.ErrorCode.INVALID_EMAIL;
 import static com.zerobase.munbanggu.type.ErrorCode.INVALID_PHONE;
-import static com.zerobase.munbanggu.type.ErrorCode.EMAIL_NOT_EXISTS;
+import static com.zerobase.munbanggu.type.ErrorCode.EMAIL_NOT_EXIST;
 import static com.zerobase.munbanggu.type.ErrorCode.USER_NOT_EXIST;
 import static com.zerobase.munbanggu.type.ErrorCode.USER_WITHDRAWN;
 import static com.zerobase.munbanggu.type.ErrorCode.WRONG_PASSWORD;
@@ -95,7 +95,7 @@ public class UserService {
     public GetUserDto getInfo(String email) {
 
         User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new UserException(EMAIL_NOT_EXISTS));
+                .orElseThrow(() -> new UserException(EMAIL_NOT_EXIST));
 
         return GetUserDto.builder().
                 email(user.getEmail())
@@ -171,7 +171,7 @@ public class UserService {
      */
     public User findByEmailAndPhone(String email, String phone) {
         User user = userRepository.findByEmail(email)
-            .orElseThrow(() -> new UserException(EMAIL_NOT_EXISTS));  //이메일이 db에 있는지 확인
+            .orElseThrow(() -> new UserException(EMAIL_NOT_EXIST));  //이메일이 db에 있는지 확인
 
         if (user.getPhone().equals(phone))
             return user;
@@ -229,7 +229,7 @@ public class UserService {
         // 핸드폰 인증 번호가 같으면
         if( info != null && info.getVerificationCode().equals(resetPwDto.getInputCode())) {
             User user = userRepository.findByEmail(info.getEmail())
-                .orElseThrow(() -> new UserException(EMAIL_NOT_EXISTS));
+                .orElseThrow(() -> new UserException(EMAIL_NOT_EXIST));
             user.setPassword(resetPwDto.getNewPassword());
             userRepository.save(user);
