--- src/main/java/com/zerobase/munbanggu/user/type/EnrollmentStatus.java
@@ -1,5 +1,0 @@
-package com.zerobase.munbanggu.user.type;
-
-public enum EnrollmentStatus {
-    OPEN, CLOSE
-}
