--- src/main/java/com/zerobase/munbanggu/user/type/RefundCycle.java
@@ -1,16 +1,0 @@
-package com.zerobase.munbanggu.user.type;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public enum RefundCycle {
-
-    ONE(1), TWO(2), THREE(3),
-    FOUR(4), FIVE(5),SIX(6);
-
-    private final int value;
-
-    public int getValue(){
-        return value;
-    }
-}
